# contracts/openapi/layer4-agents.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Layer 4: Agentic Workflow Engine",
    "description": "LangGraph-powered workflow orchestration for Value Fabric with multi-agent support",
    "version": "0.2.0"
  },
  "paths": {
...
    "/auth/oidc/{tenant_slug}/login": {
      "get": {
        "tags": [
          "OIDC SSO"
        ],
        "summary": "Oidc Login",
        "description": "Initiate OIDC login for a tenant.\n\nStores state/nonce in ``oidc_sessions`` and returns the IdP authorize URL.",
        "operationId": "oidc_login_auth_oidc__tenant_slug__login_get",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          },
          {
            "name": "redirect_uri",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Redirect Uri"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Oidc Login Auth Oidc  Tenant Slug  Login Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oidc/callback": {
      "get": {
        "tags": [
          "OIDC SSO"
        ],
        "summary": "Oidc Callback",
        "description": "Handle OIDC callback from the identity provider.\n\nValidates state, exchanges code for tokens, verifies id_token,\nmaps claims to VF role, auto-provisions the user if enabled,\nand returns an internal JWT access_token.",
        "operationId": "oidc_callback_auth_oidc_callback_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Oidc Callback Auth Oidc Callback Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/oidc/{tenant_slug}/metadata": {
      "get": {
        "tags": [
          "OIDC SSO"
        ],
        "summary": "Oidc Metadata",
        "description": "Return non-sensitive OIDC configuration for a tenant.",
        "operationId": "oidc_metadata_auth_oidc__tenant_slug__metadata_get",
        "parameters": [
          {
            "name": "tenant_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Oidc Metadata Auth Oidc  Tenant Slug  Metadata Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/oidc/{tenant_slug}/login": {
      "get": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "GET /v1/auth/oidc/{tenant_slug}/login",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\tenants\\api\\routes\\oidc.py).",
        "operationId": "discovered_get_v1_auth_oidc_tenant_slug_login",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/v1/auth/oidc/callback": {
      "get": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "GET /v1/auth/oidc/callback",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\tenants\\api\\routes\\oidc.py).",
        "operationId": "discovered_get_v1_auth_oidc_callback",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/v1/auth/oidc/{tenant_slug}/metadata": {
      "get": {
        "tags": [
          "layer4-agents"
        ],
        "summary": "GET /v1/auth/oidc/{tenant_slug}/metadata",
        "description": "Discovered from router source (value-fabric\\layer4-agents\\src\\tenants\\api\\routes\\oidc.py).",
        "operationId": "discovered_get_v1_auth_oidc_tenant_slug_metadata",
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}